+++ EntranceShuffle.py
@@ -1,7 +1,8 @@
 import logging
 from collections import defaultdict
 import RaceRandom as random
-from BaseClasses import CollectionState, RegionType, Terrain
+from BaseClasses import CollectionState, RegionType
+from OverworldShuffle import build_accessible_region_list
 from OWEdges import OWTileRegions
 
 entrance_pool = list()
@@ -28,7 +29,9 @@ def link_entrances(world, player):
     Old_Man_House = Old_Man_House_Base.copy()
     Cave_Three_Exits = Cave_Three_Exits_Base.copy()
 
-    sectors = build_sectors(world, player)
+    from OverworldShuffle import build_sectors
+    if not world.owsectors[player]:
+        world.owsectors[player] = build_sectors(world, player)
 
     # modifications to lists
     if invFlag == (0x1b in world.owswaps[player][0] and world.owMixed[player]):
@@ -241,10 +244,10 @@ def link_entrances(world, player):
 
         if invFlag:
             # place dark sanc
-            place_dark_sanc(world, sectors, player)
+            place_dark_sanc(world, player)
         
         # place links house
-        links_house = place_links_house(world, sectors, player)
+        links_house = place_links_house(world, player)
         
         # place blacksmith, has limited options
         place_blacksmith(world, links_house, player)
@@ -271,10 +274,10 @@ def link_entrances(world, player):
 
         # place dark sanc
         if invFlag:
-            place_dark_sanc(world, sectors, player)
+            place_dark_sanc(world, player)
         
         # place links house
-        links_house = place_links_house(world, sectors, player)
+        links_house = place_links_house(world, player)
         
         # place blacksmith, has limited options
         place_blacksmith(world, links_house, player)
@@ -331,10 +334,10 @@ def link_entrances(world, player):
 
         # place dark sanc
         if invFlag:
-            place_dark_sanc(world, sectors, player, list(zip(*drop_connections + dropexit_connections))[0])
+            place_dark_sanc(world, player, list(zip(*drop_connections + dropexit_connections))[0])
         
         # place links house
-        links_house = place_links_house(world, sectors, player, list(zip(*drop_connections + dropexit_connections))[0])
+        links_house = place_links_house(world, player, list(zip(*drop_connections + dropexit_connections))[0])
 
         # determine pools
         lw_entrances = list()
@@ -408,7 +411,7 @@ def link_entrances(world, player):
         scramble_holes(world, player)
 
         # place links house
-        links_house = place_links_house(world, sectors, player)
+        links_house = place_links_house(world, player)
         
         # place blacksmith, has limited options
         place_blacksmith(world, links_house, player)
@@ -494,7 +497,7 @@ def link_entrances(world, player):
         scramble_holes(world, player)
 
         # place links house
-        links_house = place_links_house(world, sectors, player)
+        links_house = place_links_house(world, player)
         
         # place blacksmith, has limited options
         place_blacksmith(world, links_house, player)
@@ -542,10 +545,10 @@ def link_entrances(world, player):
 
         # place dark sanc
         if invFlag:
-            place_dark_sanc(world, sectors, player)
+            place_dark_sanc(world, player)
         
         # place links house
-        links_house = place_links_house(world, sectors, player)
+        links_house = place_links_house(world, player)
         
         # place blacksmith, has limited options
         place_blacksmith(world, links_house, player)
@@ -616,10 +619,10 @@ def link_entrances(world, player):
 
         # place dark sanc
         if invFlag:
-            place_dark_sanc(world, sectors, player)
+            place_dark_sanc(world, player)
 
         # place links house
-        links_house = place_links_house(world, sectors, player)
+        links_house = place_links_house(world, player)
 
         # place blacksmith, place sanc exit first for additional blacksmith candidates
         doors = list(entrance_pool)
@@ -1371,7 +1374,7 @@ def full_shuffle_dungeons(world, Dungeon_Exits, player):
     connect_caves(world, lw_entrances, dw_entrances, dungeon_exits, player)
 
 
-def place_links_house(world, sectors, player, ignore_list=[]):
+def place_links_house(world, player, ignore_list=[]):
     invFlag = world.mode[player] == 'inverted'
     if world.mode[player] == 'standard' or not world.shufflelinks[player]:
         links_house = 'Links House' if not invFlag else 'Big Bomb Shop'
@@ -1383,9 +1386,9 @@ def place_links_house(world, sectors, player, ignore_list=[]):
                     break
         
         if invFlag and isinstance(dark_sanc, str):
-            links_house_doors = [i for i in get_distant_entrances(world, dark_sanc, sectors, player) if i in entrance_pool]
+            links_house_doors = [i for i in get_distant_entrances(world, dark_sanc, player) if i in entrance_pool]
         else:
-            links_house_doors = [i for i in get_starting_entrances(world, sectors, player, world.shuffle[player] != 'insanity') if i in entrance_pool]
+            links_house_doors = [i for i in get_starting_entrances(world, player, world.shuffle[player] != 'insanity') if i in entrance_pool]
         if world.shuffle[player] in ['lite', 'lean']:
             links_house_doors = [e for e in links_house_doors if e in list(zip(*(default_item_connections + (default_shop_connections if world.shopsanity[player] else []))))[0]]
         
@@ -1397,11 +1400,11 @@ def place_links_house(world, sectors, player, ignore_list=[]):
     return links_house
 
 
-def place_dark_sanc(world, sectors, player, ignore_list=[]):
+def place_dark_sanc(world, player, ignore_list=[]):
     if not world.shufflelinks[player]:
-        sanc_doors = [i for i in get_distant_entrances(world, 'Big Bomb Shop', sectors, player) if i in entrance_pool]
+        sanc_doors = [i for i in get_distant_entrances(world, 'Big Bomb Shop', player) if i in entrance_pool]
     else:
-        sanc_doors = [i for i in get_starting_entrances(world, sectors, player, world.shuffle[player] != 'insanity') if i in entrance_pool]
+        sanc_doors = [i for i in get_starting_entrances(world, player, world.shuffle[player] != 'insanity') if i in entrance_pool]
     if world.shuffle[player] in ['lite', 'lean']:
         sanc_doors = [e for e in sanc_doors if e in list(zip(*(default_item_connections + (default_shop_connections if world.shopsanity[player] else []))))[0]]
     
@@ -1634,100 +1637,6 @@ def unbias_dungeons(Dungeon_Exits):
     tuplize_lists_in_list(Dungeon_Exits)
 
 
-def build_sectors(world, player):
-    from Main import copy_world
-    from OWEdges import OWTileRegions
-    
-    # perform accessibility check on duplicate world
-    for player in range(1, world.players + 1):
-        world.key_logic[player] = {}
-    base_world = copy_world(world)
-    world.key_logic = {}
-    
-    # build lists of contiguous regions accessible with full inventory (excl portals/mirror/flute/entrances)
-    regions = list(OWTileRegions.copy().keys())
-    sectors = list()
-    while(len(regions) > 0):
-        explored_regions = build_accessible_region_list(base_world, regions[0], player, False, False, False, False)
-        regions = [r for r in regions if r not in explored_regions]
-        unique_regions = [_ for i in range(len(sectors)) for _ in sectors[i]]
-        if (any(r in unique_regions for r in explored_regions)):
-            for s in range(len(sectors)):
-                if (any(r in sectors[s] for r in explored_regions)):
-                    sectors[s] = set(list(sectors[s]) + list(explored_regions))
-                    break
-        else:
-            sectors.append(explored_regions)
-    
-    # remove water regions if Flippers not in starting inventory
-    if not any(map(lambda i: i.name == 'Flippers', world.precollected_items)):
-        for s in range(len(sectors)):
-            terrains = list()
-            for regionname in sectors[s]:
-                region = world.get_region(regionname, player)
-                if region.terrain == Terrain.Land:
-                    terrains.append(regionname)
-            sectors[s] = terrains
-    
-    # within each group, split into contiguous regions accessible only with starting inventory
-    for s in range(len(sectors)):
-        regions = list(sectors[s]).copy()
-        sectors2 = list()
-        while(len(regions) > 0):
-            explored_regions = build_accessible_region_list(base_world, regions[0], player, False, False, True, False)
-            regions = [r for r in regions if r not in explored_regions]
-            unique_regions = [_ for i in range(len(sectors2)) for _ in sectors2[i]]
-            if (any(r in unique_regions for r in explored_regions)):
-                for s2 in range(len(sectors2)):
-                    if (any(r in sectors2[s2] for r in explored_regions)):
-                        sectors2[s2] = set(list(sectors2[s2]) + list(explored_regions))
-                        break
-            else:
-                sectors2.append(explored_regions)
-        sectors[s] = sectors2
-
-    return sectors
-
-
-def build_accessible_region_list(world, start_region, player, build_copy_world=False, cross_world=False, region_rules=True, ignore_ledges = False):
-    from Main import copy_world
-    from Items import ItemFactory
-    
-    def explore_region(region_name, region=None):
-        explored_regions.add(region_name)
-        if not region:
-            region = base_world.get_region(region_name, player)
-        for exit in region.exits:
-            if exit.connected_region is not None:
-                if any(map(lambda i: i.name == 'Ocarina', base_world.precollected_items)) and exit.spot_type == 'Flute':
-                    fluteregion = exit.connected_region
-                    for flutespot in fluteregion.exits:
-                        if flutespot.connected_region and flutespot.connected_region.name not in explored_regions:
-                            explore_region(flutespot.connected_region.name, flutespot.connected_region)
-                elif exit.connected_region.name not in explored_regions \
-                        and (exit.connected_region.type == region.type or (cross_world and exit.connected_region.type in [RegionType.LightWorld, RegionType.DarkWorld])) \
-                        and (not region_rules or exit.access_rule(blank_state)) and (not ignore_ledges or exit.spot_type != 'Ledge'):
-                    explore_region(exit.connected_region.name, exit.connected_region)
-    
-    if build_copy_world:
-        for player in range(1, world.players + 1):
-            world.key_logic[player] = {}
-        base_world = copy_world(world)
-        base_world.override_bomb_check = True
-        world.key_logic = {}
-    else:
-        base_world = world
-    
-    connect_simple(base_world, 'Links House S&Q', start_region, player)
-    blank_state = CollectionState(base_world)
-    if base_world.mode[player] == 'standard':
-        blank_state.collect(ItemFactory('Zelda Delivered', player), True)
-    explored_regions = set()
-    explore_region(start_region)
-
-    return explored_regions
-    
-
 def build_accessible_entrance_list(world, start_region, player, assumed_inventory=[], cross_world=False, region_rules=True, exit_rules=True, include_one_ways=False):
     from Main import copy_world
     from Items import ItemFactory
@@ -1766,7 +1675,7 @@ def build_accessible_entrance_list(world, start_region, player, assumed_inventor
     return entrances
     
 
-def get_starting_entrances(world, sectors, player, force_starting_world=True):
+def get_starting_entrances(world, player, force_starting_world=True):
     invFlag = world.mode[player] == 'inverted'
 
     # find largest walkable sector
@@ -1775,7 +1684,7 @@ def get_starting_entrances(world, sectors, player, force_starting_world=True):
     entrances = list()
     while not len(entrances):
         while (sector is None):
-            sector = max(sectors, key=lambda x: len(x) - (0 if x not in invalid_sectors else 1000))
+            sector = max(world.owsectors[player], key=lambda x: len(x) - (0 if x not in invalid_sectors else 1000))
             if not ((world.owCrossed[player] == 'polar' and world.owMixed[player]) or world.owCrossed[player] not in ['none', 'polar']) \
                     and world.get_region(next(iter(next(iter(sector)))), player).type != (RegionType.LightWorld if not invFlag else RegionType.DarkWorld):
                 invalid_sectors.append(sector)
@@ -1799,10 +1708,10 @@ def get_starting_entrances(world, sectors, player, force_starting_world=True):
     return entrances
 
 
-def get_distant_entrances(world, start_entrance, sectors, player):
+def get_distant_entrances(world, start_entrance, player):
     # get walkable sector in which initial entrance was placed
     start_region = world.get_entrance(start_entrance, player).parent_region.name
-    regions = next(s for s in sectors if any(start_region in w for w in s))
+    regions = next(s for s in world.owsectors[player] if any(start_region in w for w in s))
     regions = next(w for w in regions if start_region in w)
     
     # eliminate regions surrounding the initial entrance until less than half of the candidate regions remain